refactor: extract WorkspaceState notifications into an extension file
Move the local-notification listener, delivery, dedup, and permission
helpers into WorkspaceState+Notifications.swift as `extension
WorkspaceState`. Also promote the file-scope private extensions in
WorkspaceState.swift (conversion helpers on NotificationSettingsSnapshot,
UNAuthorizationStatus, ProfileDraft, RelaySettingsSnapshot, KeyPackageItem)
to internal, since methods moving to other files need to see them — the same
file-scoped-access enablement as the prior commit, for top-level helpers.

Pure code movement, no behavior change. Build green; all 208 tests pass.
